+++ R/deprecated.R
@@ -1,29 +1,11 @@
-#' @include connection.R
+#' @include MySQLConnection.R
 NULL
 
 #' Make R/S-Plus identifiers into legal SQL identifiers
 #'
-#' These methods are straight-forward implementations of the corresponding
-#' generic functions.
+#' DEPRECATED: please use \code{dbQuoteIdentifier} instead.
 #'
-#' @param dbObj any MySQL object (e.g., \code{MySQLDriver}).
-#' @param snames a character vector of R/S-Plus
-#'   identifiers (symbols) from which we need to make SQL identifiers.
-#' @param name a character vector of SQL identifiers we want to check against
-#'   keywords from the DBMS.
-#' @param unique logical describing whether the resulting set of SQL names
-#'   should be unique.  Its default is \code{TRUE}. Following the SQL 92
-#'   standard, uniqueness of SQL identifiers is determined regardless of whether
-#'   letters are upper or lower case.
-#' @param allow.keywords logical describing whether SQL keywords should be
-#'   allowed in the resulting set of SQL names.  Its default is \code{TRUE}
-#' @param keywords a character vector with SQL keywords, by default it is
-#'   \code{.MySQLKeywords} define in \code{RMySQL}. This may be overriden by
-#'   users.
-#' @param case a character string specifying whether to make the
-#'   comparison as lower case, upper case, or any of the two.  it defaults to
-#'   \code{any}.
-#' @param ... Unused, needed for compatibility with generic.
+#' @keywords internal
 #' @export
 setMethod("make.db.names", c("MySQLConnection", "character"),
   function(dbObj, snames, keywords, unique, allow.keywords, ...) {
@@ -85,3 +67,103 @@ setMethod("isSQLKeyword", c("MySQLConnection", "character"),
   "VARCHARACTER", "VARYING", "WHEN", "WHERE", "WHILE", "WITH",
   "WRITE", "XOR", "YEAR_MONTH", "ZEROFILL"
 )
+
+#' Check if a database object is valid.
+#'
+#' DEPRECATED
+#'
+#' @keywords internal
+#' @return a logical scalar.
+#' @export
+#' @examples
+#' dbIsValid(MySQL())
+isIdCurrent <- function(obj)  {
+  .Deprecated("dbIsValid")
+  dbIsValid(obj)
+}
+
+
+#' Escape SQL-special characters in strings.
+#'
+#' DEPRECATED: Please use \code{dbQuoteString} instead.
+#'
+#' @export
+#' @keywords internal
+setGeneric("dbEscapeStrings", function(con, strings, ...) {
+  .Deprecated("dbQuoteString")
+  dbQuoteString(con, strings, ..)
+})
+
+#' Build the SQL CREATE TABLE definition as a string
+#'
+#' DEPRECATED
+#'
+#' @keywords internal
+#' @export
+mysqlBuildTableDefinition <- function(dbObj, name, obj, field.types = NULL,
+  row.names = TRUE, ...) {
+
+  warning("Deprecated. Please use SQL::sqlTableCreate instead.")
+
+  if (!is.data.frame(obj)) {
+    obj <- as.data.frame(obj)
+  }
+  value <- explict_rownames(obj, row.names)
+
+  if (is.null(field.types)) {
+    field.types <- vapply(value, dbDataType, dbObj = dbObj,
+      FUN.VALUE = character(1))
+  }
+  # Escape field names
+  names(field.types) <- dbQuoteIdentifier(dbObj, names(field.types))
+
+  flds <- paste(names(field.types), field.types)
+  paste("CREATE TABLE", name, "\n(", paste(flds, collapse = ",\n\t"), "\n)")
+}
+
+## Escape problematic characters in the data frame.
+## These are: - tab, as this is the field separator
+##            - newline, as this is the record separator
+##            - backslash, the escaping character
+## Obviously, not all data types can contain these, e.g. numeric types
+## can not. So we only substitute character and factor types.
+## (FIXME: is there anything else?)
+escape <- function(table) {
+  table <- as.data.frame(table)
+  repcols <- which(sapply(table, is.character) | sapply(table, is.factor))
+  for (rc in repcols) {
+    table[,rc] <- gsub("\\\\", "\\\\\\\\", table[,rc])
+    table[,rc] <- gsub("\\n", "\\\\n", table[,rc])
+    table[,rc] <- gsub("\\t", "\\\\t", table[,rc])
+  }
+  table
+}
+
+
+## safe.write makes sure write.table doesn't exceed available memory by batching
+## at most batch rows (but it is still slowww)
+safe.write <- function(value, file, batch, ...) {
+  N <- nrow(value)
+  if(N<1){
+    warning("no rows in data.frame")
+    return(NULL)
+  }
+  digits <- options(digits = 17)
+  on.exit(options(digits))
+  if(missing(batch) || is.null(batch))
+    batch <- 10000
+  else if(batch<=0)
+    batch <- N
+  from <- 1
+  to <- min(batch, N)
+  conb <- file(file,open="wb")
+  while(from<=N){
+    write.table(escape(value[from:to,, drop=FALSE]), file = conb,
+      append = TRUE, quote = FALSE, sep="\t", na = "\\N",
+      row.names=FALSE, col.names=FALSE, eol = '\n', ...)
+    from <- to+1
+    to <- min(to+batch, N)
+  }
+  close(conb)
+  invisible(NULL)
+}